Tickets' messages deletion done Beggining FAQ

+++ app/controllers/_DefaultController.php
@@ -24,29 +24,34 @@ class _DefaultController extends \BaseController {
 	 * @see BaseController::index()
 	 */
 	public function index($message=null){
-		global $config;
-		$baseHref=get_class($this);
-		if(isset($message)){
-			if(is_string($message)){
-				$message=new DisplayedMessage($message);
-			}
-			$message->setTimerInterval($this->messageTimerInterval);
-			$this->_showDisplayedMessage($message);
-		}
-		$objects=DAO::getAll($this->model);
-		echo "<table class='table table-striped'>";
-		echo "<thead><tr><th>".$this->model."</th></tr></thead>";
-		echo "<tbody>";
-		foreach ($objects as $object){
-			echo "<tr>";
-			echo "<td>".$object->toString()."</td>";
-			echo "<td class='td-center'><a class='btn btn-primary btn-xs' href='".$baseHref."/frm/".$object->getId()."'><span class='glyphicon glyphicon-edit' aria-hidden='true'></span></a></td>".
-			"<td class='td-center'><a class='btn btn-warning btn-xs' href='".$baseHref."/delete/".$object->getId()."'><span class='glyphicon glyphicon-remove' aria-hidden='true'></span></a></td>";
-			echo "</tr>";
-		}
-		echo "</tbody>";
-		echo "</table>";
-		echo "<a class='btn btn-primary' href='".$config["siteUrl"].$baseHref."/frm'>Ajouter...</a>";
+        if(Auth::isAdmin()) {
+            global $config;
+            $baseHref = get_class($this);
+            if (isset($message)) {
+                if (is_string($message)) {
+                    $message = new DisplayedMessage($message);
+                }
+                $message->setTimerInterval($this->messageTimerInterval);
+                $this->_showDisplayedMessage($message);
+            }
+            $objects = DAO::getAll($this->model);
+            echo "<table class='table table-striped'>";
+            echo "<thead><tr><th>" . $this->model . "</th></tr></thead>";
+            echo "<tbody>";
+            foreach ($objects as $object) {
+                echo "<tr>";
+                echo "<td>" . $object->toString() . "</td>";
+                echo "<td class='td-center'><a class='btn btn-primary btn-xs' href='" . $baseHref . "/frm/" . $object->getId() . "'><span class='glyphicon glyphicon-edit' aria-hidden='true'></span></a></td>" .
+                    "<td class='td-center'><a class='btn btn-warning btn-xs' href='" . $baseHref . "/delete/" . $object->getId() . "'><span class='glyphicon glyphicon-remove' aria-hidden='true'></span></a></td>";
+                echo "</tr>";
+            }
+            echo "</tbody>";
+            echo "</table>";
+            echo "<a class='btn btn-primary' href='" . $config["siteUrl"] . $baseHref . "/frm'>Ajouter...</a>";
+        }
+        else {
+            $this->messageDanger("Vous n'avez pas l'autorisation d'accéder à cette page.");
+        }
 	}
 
 	/**
